requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.5
django-filter==22.1
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
promise==2.3
psycopg2==2.9.5
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
